Move index.tsx to (index) subfolder and move all components there too
