+++ application/account-management/WebApp/pages/(index)/-components/FooterSection.tsx
