+++ application/account-management/WebApp/pages/(index)/-components/HeroSection.tsx
