+++ application/account-management/WebApp/pages/(index)/-components/TechnologySection.tsx
