+++ application/account-management/WebApp/pages/(index)/-components/TechnologySection2.tsx
